successive bye fitness
added logic to calculated the plays vs time
which is then used to create a fitness criteria
for no team to have two byes in a row. This commit also
includes a small change to pull the bye requirement
for each division from the facility structure instead
of a hardcoded list.