Adding tool for removing last names from roster
Adding a simple tool that takes in a csv path and creates a new csv
with the last names sanitized. This means leaving as few letters
in the last name as possible while still uniquely identifying each
person in each division.

There are some bugs in here, like:
-people with the same name, differentialted by jr and sr, will have
their last names included
-names htat contain other names (dave struck, dave strucker) will
randomly be replaced to one of hte two, but this has not occured yet.
These are not critical issue.

Also, there is also some dead code from a previous draft that I'm
storing to git here before deleting it.