Add docs; add open play report, add switch tool
First off, I added more context to the existing debug reports so I could
send it to other people and have it be reasonably comprehensible. This
stuff works but all of the output-list join stuff really needs to be
cleaned up at some point.

I added the open play evenness metric in fitness and needed this
new report to assess if it was working correctly.

Added some more logic for switching teams and reffing for ocmpleted
schedule to aide customer work needed to support skills clinics
and other odd needs that are hard to automated.