addding get openplay times logic
the existing scheduling logic required required the ability
to get the list of all open courts that need to be scheduled.
I added a simple routine for collecting these and collapsing
them into a list of times to support the need in the fitness
function to assess which teams have the opertunities to make
use of open court time.