non-functional attempt a google sheet access
I'm working remove the need for csvs in our workflow.
the answer to that is using the googledrive / google sheets apis
to pipe data between steps. This file is directly drawn from
google sample pages, but is not functional yet for my purposes.
I'm uploading here as I'm losing this computer.