Adding interface for updating season teams
This script uses the csv read and team name cleaner to get
a set of anonymized team member names, then uploads
them to the scvl database. Authentication is stored in
an environment variable. There are parts of this that
I simply hard-coded for this first pass. Specifically
the season and division numbers, which I created
with sql in teh database itself. These steps could
easily be automated in the next iteration.

In addition to that, I think refactoring the database
connection into a function or wrapper could simplify
things a bit.