Ref Flexability; hard-code corrections; team formatting
To many changes for one commit but was sick when I did most of this.

Added some logic to make things not break if reffing is a None or a
(div, team) tuple.

Fixed some hard-coded court and division values
from back in the SCVL days. They are all based on Facilities as they
should be now.

I made a minor change in how games are formatted to right justify
single digit values. This makes reviewing ascii outputs easier.