Add draft add_sand_reffing; fix division hard-coding
I fixed another location of division being hard-coded 5, to be
based on the team_counts specified as input.

I added draft logic for the concept of add_sand_reffing. I was,
to note, sick when I had this notition. It would involve allowing
inter-division reffing which solved some other problems I was
having. Howerver, it cut rather severily againt the division/day
fitness logic on which... all this is based. I realized it was
easier to fix the blocking so that all division could cover than own
then do this. However, this could be a thing some day. So I'm saving
it here before deleting it from the main line.