Adding a module for caching and recalling schedules
I made some utilities for easily saving models to pickle
files and back, and bundled it in the optimizer.py
module (which is way too big).

I'm making a new module to house all of that logic. For
now, I'm just adding utility program on top of things
that I can use for rapidly developing my sand schedule
formatter for sheets export.