Initial structure of reffing splitting
In Stonewall sand, reffing is a two-part activity where
each reffing team provides two-line judges to 1 court
and an up-ref to another.

I'm adding post-processing logic to the solve to take the
SCVL compatible, single-ref structure and adjust it to
track both these refs. This is final transformation step
before the score gets flattly formatted and pushed to the
output trix.