Add working schedule upload function
I finished off the stub routine for uploading
a schedule to the target sheet in the config.
As expected, I found a few bugs in my schedule
formatting along the way.

Note: I'm semi-hard cording the court options
here to 4 or 5. However, when this fails, it
will fail quite explicitly.