Adding deep copy to avoid schdule corruption
The last set in the sand schedule exporting process
is to split out the up refs from teh lines refs. This
changes the refs from an int (division defined by game.div)
to a tup of tups (up then line). By adding in this deep copy
the schedule spliting process doesn't corrupt the initial
schedule, which was causing the audit reporting to fail.