Adding "No Reffing" Message to teamwise schedule
Added a routine to check if we saw reffing in someones
schedule and, if not, add a statement saying there is
no reffing. Note, the flags I used here not bullet proof.
Specifically, I think if there is a team name with ref
in it that could cause an issue. With week, same issue
in the other direction that it is tightly coupled to my
day headers in a separate function. Still, its low stake
as we are "risking" someone with reffing in the schedule
... not having a statement saying they don't have reffing
-> done for today.