Final touchups on csv exporting: arg driven, refs
I fixed the hours, which I had off by one. Need to
make this dynamic from the template in the next
iteration.

Adding the line reffing duties in addition to the
up reffing duties.

Change the local run to be arg driven, so I can update the
schedule being pushed into it.