String report for faclities
I want the facilities object to encode a variety of different formats in a consistent internal format. I may want an adult test at some point, but for not, a human readibly export of the content that I can diff vs a known (and visual review myself), is a fast test with season apporach.

note:t his is not working exactly right yet. I want a dump of all the game contents as a table.