facility variables and constaint iterables
adding the concept of applying facilities constaints to the model. I thikn I'll need to do this in a loop for all the games (which I don't thikn exiect currenlty), but am starting out with using it as a way to launder the games_per_season into the constriants without needing htem to understand the facilities ojbect.