Adding rec in lower courts post processor
so, I now know that I can't just alter the soruce data structures on this. I think updating the reports is a reasonable apprach, but I need to be diligent that all of my downstream interfaces (debug schedule, googel sheets upload) interface with the reports and not the model. I may do that next.