Realized I need to scale non timevariety logic by 100x
I forgot that, do quasi deal with floats, the time variety optimation is multiplying its error functions by 100. This meant it was overpowering the important triple action and by week optimizations. So, sacling the critical ones. Need to go back and structure this to be more intutive at some point, maybe seeing if we can minimize on float values instead of ints.