Adding callback for status updates
this was my last iteration before the final schedule for 2025. In it, I added a status updater while the system is running. I also, also extended the run time and hard-coded it to fail if optimization is greater than 80k. I'll take that out once hte schedule is published. It is make sure I don't accidently make the schedule worse.