Transforming three our checks
In my old system, I was unable to entirely get rid of 3 hours days. However, when I implemented it here as an optimizer, it could do it every time.

Then, we had to scrap my first schedule as it has a double header for a team at 1pm and 4pm. Because that, I decide to just translate the optimization into a constraint, simply saying, if you play or ref (busy) at X, you are not doing any at X-2, X+2, X-3, etc.