line refs and objective targets
first, for league apps to know who was line reffing, I needed to change the parser to pull that data from the uploaded schedule (I haven't made a persistent local cache yet).

Second, I change the system to store the best score for a schedule with it in the google sheet, so that it could be synced between runs (and then made the final schedule in one run).