Reworking for line ref parsing
When I made this the first time, I didn't parse line refs (just up). When I added line ref parsing in gsheets to support league apps, this broke. I gave cursor guidance on how to know when it had fixed this, and it looped a bunch of times, and then created this.